Fixx blockreward in api

+++ src/Cybercore/Blockchain/Abstractions.cs
@@ -12,7 +12,7 @@ namespace Cybercore.Blockchain
         public string NextNetworkBits { get; set; }
         public DateTime? LastNetworkBlockTime { get; set; }
         public ulong BlockHeight { get; set; }
-        public long BlockReward { get; set; }
+        public double BlockReward { get; set; }
         public int ConnectedPeers { get; set; }
         public string RewardType { get; set; }
     }
